Maybe silence swiftlint
